fix(sdk): autofix and enforce flake8-simplify (SIM) rules (#11295)
Description
-----------
<!--
Include reference to internal ticket "Fixes WB-NNNNN" and/or GitHub issue "Fixes #NNNN" (if applicable)
-->
- Fixes WB-NNNNN
- Fixes #NNNN

There are a lot of minor but repeated/cumulative code-quality issues that are readily improved (and auto-fixed + avoided going forward) by enforcing existing flake8-simplify (SIM*) rules.

This PR does that.

<!--
NEW: We're using a new changelog format that's more useful for users. Please
see CHANGELOG.unreleased.md for details and update on relevant changes such as feature
additions, bug fixes, or removals/deprecations.
-->
- [x] I updated CHANGELOG.unreleased.md, or it's not applicable


Testing
-------
How was this PR tested?

<!--
Ensure PR title compliance with the [conventional commits standards](https://github.com/wandb/wandb/blob/main/CONTRIBUTING.md#conventional-commits)
-->

+++ wandb/sdk/wandb_run.py
@@ -297,12 +297,11 @@ class RunStatusChecker:
             from wandb.agents import pyagent
 
             stop_status = result.response.stop_status_response
-            if stop_status.run_should_stop:
+            if stop_status.run_should_stop and not pyagent.is_running():  # type: ignore
                 # TODO(frz): This check is required
                 # until WB-3606 is resolved on server side.
-                if not pyagent.is_running():  # type: ignore
-                    interrupt.interrupt_main()
-                    return
+                interrupt.interrupt_main()
+                return
 
         with wb_logging.log_to_run(self._run_id):
             try:
@@ -1744,7 +1743,7 @@ class Run:
         if not isinstance(data, Mapping):
             raise TypeError("wandb.log must be passed a dictionary")
 
-        if any(not isinstance(key, str) for key in data.keys()):
+        if any(not isinstance(key, str) for key in data):
             raise TypeError("Key values passed to `wandb.log` must be strings.")
 
         self._partial_history_callback(data, step, commit)
@@ -2495,9 +2494,8 @@ class Run:
         # This is used by the "auto" resume mode, which resumes from the last
         # failed (or unfinished/crashed) run. If we reach this line, then this
         # run shouldn't be a candidate for "auto" resume.
-        if exit_code == 0:
-            if os.path.exists(self._settings.resume_fname):
-                os.remove(self._settings.resume_fname)
+        if exit_code == 0 and os.path.exists(self._settings.resume_fname):
+            os.remove(self._settings.resume_fname)
 
         try:
             self._on_finish()
@@ -2540,14 +2538,17 @@ class Run:
         if self._settings.save_code and self._settings.code_dir is not None:
             self.log_code(self._settings.code_dir)
 
-        if self._settings.x_save_requirements:
-            if self._backend and self._backend.interface:
-                from wandb.util import working_set
+        if (
+            self._settings.x_save_requirements
+            and self._backend
+            and self._backend.interface
+        ):
+            from wandb.util import working_set
 
-                logger.debug(
-                    "Saving list of pip packages installed into the current environment"
-                )
-                self._backend.interface.publish_python_packages(working_set())
+            logger.debug(
+                "Saving list of pip packages installed into the current environment"
+            )
+            self._backend.interface.publish_python_packages(working_set())
 
         if self._backend and self._backend.interface and not self._settings._offline:
             assert self._settings.run_id
@@ -2614,9 +2615,8 @@ class Run:
         # object is about to be returned to the user, don't let them modify it
         self._freeze()
 
-        if not self._settings.resume:
-            if os.path.exists(self._settings.resume_fname):
-                os.remove(self._settings.resume_fname)
+        if (not self._settings.resume) and os.path.exists(self._settings.resume_fname):
+            os.remove(self._settings.resume_fname)
 
     def _detect_and_apply_job_inputs(self) -> None:
         """If the user has staged launch inputs, apply them to the run."""
@@ -3621,7 +3621,7 @@ class Run:
             raise ValueError("level must be one of 'INFO', 'WARN', or 'ERROR'")
 
         wait_duration = wait_duration or timedelta(minutes=1)
-        if isinstance(wait_duration, int) or isinstance(wait_duration, float):
+        if isinstance(wait_duration, (int, float)):
             wait_duration = timedelta(seconds=wait_duration)
         elif not callable(getattr(wait_duration, "total_seconds", None)):
             raise TypeError(
@@ -4051,9 +4051,8 @@ def restore(
             raise ValueError(
                 "run_path required when calling wandb.restore before wandb.init"
             )
-    if root is None:
-        if run is not None:
-            root = run.dir
+    if root is None and run is not None:
+        root = run.dir
     api = public.Api()
     api_run = api.run(run_path)
     if root is None:
